chore: rename to transaction sponsorship
